Update code architecture + add other changes
Update code to use more of an MVC pattern. Still needs additional
changes to be more proper MVC code though.

Add time zone to DateTimeView, since on the production server the time
is a couple of hours behind the time on the local server.

Update environment file to use putenv instead of being a class with
static field values. This way, the environmental variables actually are
environmental variables, rather than a set of constants.

Add ENVIRONMENT environment variable, in order to produce error messages
only on the development server.